Add core remote fallback tests and inject override
