+++ crates/cli/src/lib.rs
@@ -76,17 +76,11 @@ use std::fs::{self, File};
 use std::io::{self, BufRead, BufReader, Read, Write};
 use std::num::{IntErrorKind, NonZeroU8, NonZeroU64};
 use std::path::{Path, PathBuf};
-use std::process::{Child, Command, Stdio};
 use std::str::FromStr;
-use std::sync::mpsc::{self, Sender};
-use std::thread;
 use std::time::{Duration, SystemTime};
 
 #[cfg(unix)]
-use std::os::unix::{
-    ffi::{OsStrExt, OsStringExt},
-    fs::PermissionsExt,
-};
+use std::os::unix::{ffi::OsStringExt, fs::PermissionsExt};
 
 use clap::{Arg, ArgAction, Command as ClapCommand, builder::OsStringValueParser};
 use rsync_compress::zlib::CompressionLevel;
@@ -96,8 +90,9 @@ use rsync_core::{
         BandwidthLimit, ClientConfig, ClientEntryKind, ClientEntryMetadata, ClientEvent,
         ClientEventKind, ClientProgressObserver, ClientProgressUpdate, ClientSummary,
         CompressionSetting, DirMergeEnforcedKind, DirMergeOptions, FilterRuleKind, FilterRuleSpec,
-        ModuleListRequest, TransferTimeout,
+        ModuleListRequest, RemoteFallbackArgs, TransferTimeout,
         run_client_with_observer as run_core_client_with_observer, run_module_list_with_password,
+        run_remote_transfer_fallback,
     },
     message::{Message, Role},
     rsync_error,
@@ -105,7 +100,6 @@ use rsync_core::{
 };
 use rsync_logging::MessageSink;
 use rsync_protocol::{ParseProtocolVersionErrorKind, ProtocolVersion};
-use tempfile::NamedTempFile;
 use time::{OffsetDateTime, format_description::FormatItem, macros::format_description};
 
 /// Maximum exit code representable by a Unix process.
@@ -1719,13 +1713,27 @@ where
             timeout: timeout_setting,
             out_format: out_format.clone(),
             no_motd,
+            fallback_binary: None,
             remainder,
             #[cfg(feature = "acl")]
             acls,
             #[cfg(feature = "xattr")]
             xattrs,
         };
-        return run_remote_transfer_fallback(stdout, stderr, fallback_args);
+        let fallback_result = {
+            let mut stderr_writer = stderr.writer_mut();
+            run_remote_transfer_fallback(stdout, &mut stderr_writer, fallback_args)
+        };
+        return match fallback_result {
+            Ok(code) => code,
+            Err(error) => {
+                if write_message(error.message(), stderr).is_err() {
+                    let fallback = error.message().to_string();
+                    let _ = writeln!(stderr.writer_mut(), "{}", fallback);
+                }
+                error.exit_code()
+            }
+        };
     }
 
     let numeric_ids = numeric_ids_option.unwrap_or(false);
@@ -3469,578 +3477,6 @@ fn operand_is_remote(path: &OsStr) -> bool {
     }
 
     false
-}
-
-struct RemoteFallbackArgs {
-    dry_run: bool,
-    list_only: bool,
-    archive: bool,
-    delete: bool,
-    delete_after: bool,
-    delete_excluded: bool,
-    checksum: bool,
-    size_only: bool,
-    ignore_existing: bool,
-    compress: bool,
-    compress_disabled: bool,
-    compress_level: Option<OsString>,
-    owner: Option<bool>,
-    group: Option<bool>,
-    perms: Option<bool>,
-    times: Option<bool>,
-    numeric_ids: Option<bool>,
-    sparse: Option<bool>,
-    devices: Option<bool>,
-    specials: Option<bool>,
-    relative: Option<bool>,
-    implied_dirs: Option<bool>,
-    verbosity: u8,
-    progress: bool,
-    stats: bool,
-    partial: bool,
-    remove_source_files: bool,
-    inplace: Option<bool>,
-    msgs_to_stderr: bool,
-    whole_file: Option<bool>,
-    bwlimit: Option<OsString>,
-    excludes: Vec<OsString>,
-    includes: Vec<OsString>,
-    exclude_from: Vec<OsString>,
-    include_from: Vec<OsString>,
-    filters: Vec<OsString>,
-    files_from_used: bool,
-    file_list_entries: Vec<OsString>,
-    from0: bool,
-    password_file: Option<PathBuf>,
-    protocol: Option<ProtocolVersion>,
-    timeout: TransferTimeout,
-    out_format: Option<OsString>,
-    no_motd: bool,
-    remainder: Vec<OsString>,
-    #[cfg(feature = "acl")]
-    acls: Option<bool>,
-    #[cfg(feature = "xattr")]
-    xattrs: Option<bool>,
-}
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-enum FallbackStreamKind {
-    Stdout,
-    Stderr,
-}
-
-enum FallbackStreamMessage {
-    Data(FallbackStreamKind, Vec<u8>),
-    Error(FallbackStreamKind, io::Error),
-    Finished(FallbackStreamKind),
-}
-
-fn run_remote_transfer_fallback<Out, Err>(
-    stdout: &mut Out,
-    stderr: &mut MessageSink<Err>,
-    args: RemoteFallbackArgs,
-) -> i32
-where
-    Out: Write,
-    Err: Write,
-{
-    let RemoteFallbackArgs {
-        dry_run,
-        list_only,
-        archive,
-        delete,
-        delete_after,
-        delete_excluded,
-        checksum,
-        size_only,
-        ignore_existing,
-        compress,
-        compress_disabled,
-        compress_level,
-        owner,
-        group,
-        perms,
-        times,
-        numeric_ids,
-        sparse,
-        devices,
-        specials,
-        relative,
-        implied_dirs,
-        verbosity,
-        progress,
-        stats,
-        partial,
-        remove_source_files,
-        inplace,
-        msgs_to_stderr,
-        whole_file,
-        bwlimit,
-        excludes,
-        includes,
-        exclude_from,
-        include_from,
-        filters,
-        files_from_used,
-        file_list_entries,
-        from0,
-        password_file,
-        protocol,
-        timeout,
-        out_format,
-        no_motd,
-        mut remainder,
-        #[cfg(feature = "acl")]
-        acls,
-        #[cfg(feature = "xattr")]
-        xattrs,
-    } = args;
-
-    let mut command_args = Vec::new();
-    if archive {
-        command_args.push(OsString::from("-a"));
-    }
-    if dry_run {
-        command_args.push(OsString::from("--dry-run"));
-    }
-    if list_only {
-        command_args.push(OsString::from("--list-only"));
-    }
-    if delete {
-        command_args.push(OsString::from("--delete"));
-    }
-    if delete_after {
-        command_args.push(OsString::from("--delete-after"));
-    }
-    if delete_excluded {
-        command_args.push(OsString::from("--delete-excluded"));
-    }
-    if checksum {
-        command_args.push(OsString::from("--checksum"));
-    }
-    if size_only {
-        command_args.push(OsString::from("--size-only"));
-    }
-    if ignore_existing {
-        command_args.push(OsString::from("--ignore-existing"));
-    }
-    if compress {
-        command_args.push(OsString::from("--compress"));
-    } else if compress_disabled {
-        command_args.push(OsString::from("--no-compress"));
-        if whole_file.is_none() {
-            command_args.push(OsString::from("--no-whole-file"));
-        }
-    }
-    if let Some(level) = compress_level {
-        command_args.push(OsString::from("--compress-level"));
-        command_args.push(level);
-    }
-
-    push_toggle(&mut command_args, "--owner", "--no-owner", owner);
-    push_toggle(&mut command_args, "--group", "--no-group", group);
-    push_toggle(&mut command_args, "--perms", "--no-perms", perms);
-    push_toggle(&mut command_args, "--times", "--no-times", times);
-    push_toggle(
-        &mut command_args,
-        "--numeric-ids",
-        "--no-numeric-ids",
-        numeric_ids,
-    );
-    push_toggle(&mut command_args, "--sparse", "--no-sparse", sparse);
-    push_toggle(&mut command_args, "--devices", "--no-devices", devices);
-    push_toggle(&mut command_args, "--specials", "--no-specials", specials);
-    push_toggle(&mut command_args, "--relative", "--no-relative", relative);
-    push_toggle(
-        &mut command_args,
-        "--implied-dirs",
-        "--no-implied-dirs",
-        implied_dirs,
-    );
-    push_toggle(&mut command_args, "--inplace", "--no-inplace", inplace);
-    #[cfg(feature = "acl")]
-    push_toggle(&mut command_args, "--acls", "--no-acls", acls);
-    push_toggle(
-        &mut command_args,
-        "--whole-file",
-        "--no-whole-file",
-        whole_file,
-    );
-    #[cfg(feature = "xattr")]
-    push_toggle(&mut command_args, "--xattrs", "--no-xattrs", xattrs);
-
-    for _ in 0..verbosity {
-        command_args.push(OsString::from("-v"));
-    }
-    if progress {
-        command_args.push(OsString::from("--progress"));
-    }
-    if stats {
-        command_args.push(OsString::from("--stats"));
-    }
-    if partial {
-        command_args.push(OsString::from("--partial"));
-    }
-    if remove_source_files {
-        command_args.push(OsString::from("--remove-source-files"));
-    }
-    if msgs_to_stderr {
-        command_args.push(OsString::from("--msgs2stderr"));
-    }
-
-    if let Some(limit) = bwlimit {
-        command_args.push(OsString::from("--bwlimit"));
-        command_args.push(limit);
-    }
-
-    if let Some(format) = out_format {
-        command_args.push(OsString::from("--out-format"));
-        command_args.push(format);
-    }
-
-    for exclude in excludes {
-        command_args.push(OsString::from("--exclude"));
-        command_args.push(exclude);
-    }
-    for include in includes {
-        command_args.push(OsString::from("--include"));
-        command_args.push(include);
-    }
-    for path in exclude_from {
-        command_args.push(OsString::from("--exclude-from"));
-        command_args.push(path);
-    }
-    for path in include_from {
-        command_args.push(OsString::from("--include-from"));
-        command_args.push(path);
-    }
-    for filter in filters {
-        command_args.push(OsString::from("--filter"));
-        command_args.push(filter);
-    }
-
-    let files_from_temp = match prepare_file_list(&file_list_entries, files_from_used, from0) {
-        Ok(temp) => temp,
-        Err(error) => {
-            let message = rsync_error!(
-                1,
-                format!(
-                    "failed to prepare file list for fallback rsync invocation: {}",
-                    error
-                )
-            )
-            .with_role(Role::Client);
-            if write_message(&message, stderr).is_err() {
-                let _ = writeln!(
-                    stderr.writer_mut(),
-                    "failed to prepare file list for fallback rsync invocation: {}",
-                    error
-                );
-            }
-            return 1;
-        }
-    };
-
-    if let Some(temp) = files_from_temp.as_ref() {
-        command_args.push(OsString::from("--files-from"));
-        command_args.push(temp.path().as_os_str().to_os_string());
-        if from0 {
-            command_args.push(OsString::from("--from0"));
-        }
-    }
-
-    if let Some(path) = password_file {
-        command_args.push(OsString::from("--password-file"));
-        command_args.push(path.into_os_string());
-    }
-
-    if let Some(protocol) = protocol {
-        command_args.push(OsString::from("--protocol"));
-        command_args.push(OsString::from(protocol.to_string()));
-    }
-
-    match timeout {
-        TransferTimeout::Default => {}
-        TransferTimeout::Disabled => {
-            command_args.push(OsString::from("--timeout"));
-            command_args.push(OsString::from("0"));
-        }
-        TransferTimeout::Seconds(value) => {
-            command_args.push(OsString::from("--timeout"));
-            command_args.push(OsString::from(value.get().to_string()));
-        }
-    }
-
-    if no_motd {
-        command_args.push(OsString::from("--no-motd"));
-    }
-
-    command_args.append(&mut remainder);
-
-    let binary = env::var_os("OC_RSYNC_FALLBACK")
-        .filter(|value| !value.is_empty())
-        .unwrap_or_else(|| OsString::from("rsync"));
-
-    let mut command = Command::new(&binary);
-    command.args(&command_args);
-    command.stdin(Stdio::inherit());
-    command.stdout(Stdio::piped());
-    command.stderr(Stdio::piped());
-
-    let mut child = match command.spawn() {
-        Ok(child) => child,
-        Err(error) => {
-            let message = rsync_error!(
-                1,
-                format!(
-                    "failed to launch fallback rsync binary '{}': {}",
-                    Path::new(&binary).display(),
-                    error
-                )
-            )
-            .with_role(Role::Client);
-            if write_message(&message, stderr).is_err() {
-                let _ = writeln!(
-                    stderr.writer_mut(),
-                    "failed to launch fallback rsync binary '{}': {}",
-                    Path::new(&binary).display(),
-                    error
-                );
-            }
-            return 1;
-        }
-    };
-
-    let (sender, receiver) = mpsc::channel();
-    let mut stdout_thread = child
-        .stdout
-        .take()
-        .map(|handle| spawn_fallback_reader(handle, FallbackStreamKind::Stdout, sender.clone()));
-    let mut stderr_thread = child
-        .stderr
-        .take()
-        .map(|handle| spawn_fallback_reader(handle, FallbackStreamKind::Stderr, sender.clone()));
-    drop(sender);
-
-    let mut stdout_open = stdout_thread.is_some();
-    let mut stderr_open = stderr_thread.is_some();
-
-    while stdout_open || stderr_open {
-        match receiver.recv() {
-            Ok(FallbackStreamMessage::Data(FallbackStreamKind::Stdout, data)) => {
-                if let Err(error) = stdout.write_all(&data) {
-                    let message =
-                        rsync_error!(1, format!("failed to forward fallback stdout: {error}"))
-                            .with_role(Role::Client);
-                    if write_message(&message, stderr).is_err() {
-                        let _ = writeln!(
-                            stderr.writer_mut(),
-                            "failed to forward fallback stdout: {error}"
-                        );
-                    }
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return 1;
-                }
-            }
-            Ok(FallbackStreamMessage::Data(FallbackStreamKind::Stderr, data)) => {
-                if let Err(error) = stderr.writer_mut().write_all(&data) {
-                    let _ = writeln!(
-                        stderr.writer_mut(),
-                        "failed to forward fallback stderr: {error}"
-                    );
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return 1;
-                }
-            }
-            Ok(FallbackStreamMessage::Error(FallbackStreamKind::Stdout, error)) => {
-                let message = rsync_error!(
-                    1,
-                    format!("failed to read stdout from fallback rsync: {error}")
-                )
-                .with_role(Role::Client);
-                if write_message(&message, stderr).is_err() {
-                    let _ = writeln!(
-                        stderr.writer_mut(),
-                        "failed to read stdout from fallback rsync: {error}"
-                    );
-                }
-                terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                return 1;
-            }
-            Ok(FallbackStreamMessage::Error(FallbackStreamKind::Stderr, error)) => {
-                let message = rsync_error!(
-                    1,
-                    format!("failed to read stderr from fallback rsync: {error}")
-                )
-                .with_role(Role::Client);
-                if write_message(&message, stderr).is_err() {
-                    let _ = writeln!(
-                        stderr.writer_mut(),
-                        "failed to read stderr from fallback rsync: {error}"
-                    );
-                }
-                terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                return 1;
-            }
-            Ok(FallbackStreamMessage::Finished(FallbackStreamKind::Stdout)) => {
-                stdout_open = false;
-            }
-            Ok(FallbackStreamMessage::Finished(FallbackStreamKind::Stderr)) => {
-                stderr_open = false;
-            }
-            Err(_) => {
-                if stdout_open {
-                    let message =
-                        rsync_error!(1, "failed to capture stdout from fallback rsync binary")
-                            .with_role(Role::Client);
-                    if write_message(&message, stderr).is_err() {
-                        let _ = writeln!(
-                            stderr.writer_mut(),
-                            "failed to capture stdout from fallback rsync binary"
-                        );
-                    }
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return 1;
-                }
-                if stderr_open {
-                    let message =
-                        rsync_error!(1, "failed to capture stderr from fallback rsync binary")
-                            .with_role(Role::Client);
-                    if write_message(&message, stderr).is_err() {
-                        let _ = writeln!(
-                            stderr.writer_mut(),
-                            "failed to capture stderr from fallback rsync binary"
-                        );
-                    }
-                    terminate_fallback_process(&mut child, &mut stdout_thread, &mut stderr_thread);
-                    return 1;
-                }
-                break;
-            }
-        }
-    }
-
-    join_fallback_thread(&mut stdout_thread);
-    join_fallback_thread(&mut stderr_thread);
-
-    let status = match child.wait() {
-        Ok(status) => status,
-        Err(error) => {
-            let message = rsync_error!(
-                1,
-                format!("failed to wait for fallback rsync process: {error}")
-            )
-            .with_role(Role::Client);
-            if write_message(&message, stderr).is_err() {
-                let _ = writeln!(
-                    stderr.writer_mut(),
-                    "failed to wait for fallback rsync process: {error}"
-                );
-            }
-            return 1;
-        }
-    };
-
-    // Ensure the temporary file, if any, is kept alive until the end of the transfer.
-    drop(files_from_temp);
-
-    status.code().unwrap_or(MAX_EXIT_CODE)
-}
-
-fn spawn_fallback_reader<R>(
-    mut reader: R,
-    kind: FallbackStreamKind,
-    sender: Sender<FallbackStreamMessage>,
-) -> thread::JoinHandle<()>
-where
-    R: Read + Send + 'static,
-{
-    thread::spawn(move || {
-        let mut buffer = [0u8; 8192];
-        loop {
-            match reader.read(&mut buffer) {
-                Ok(0) => {
-                    let _ = sender.send(FallbackStreamMessage::Finished(kind));
-                    break;
-                }
-                Ok(n) => {
-                    if sender
-                        .send(FallbackStreamMessage::Data(kind, Vec::from(&buffer[..n])))
-                        .is_err()
-                    {
-                        break;
-                    }
-                }
-                Err(error) => {
-                    let _ = sender.send(FallbackStreamMessage::Error(kind, error));
-                    break;
-                }
-            }
-        }
-    })
-}
-
-fn join_fallback_thread(handle: &mut Option<thread::JoinHandle<()>>) {
-    if let Some(join_handle) = handle.take() {
-        let _ = join_handle.join();
-    }
-}
-
-fn terminate_fallback_process(
-    child: &mut Child,
-    stdout_thread: &mut Option<thread::JoinHandle<()>>,
-    stderr_thread: &mut Option<thread::JoinHandle<()>>,
-) {
-    let _ = child.kill();
-    let _ = child.wait();
-    join_fallback_thread(stdout_thread);
-    join_fallback_thread(stderr_thread);
-}
-
-fn prepare_file_list(
-    entries: &[OsString],
-    files_from_used: bool,
-    zero_terminated: bool,
-) -> io::Result<Option<NamedTempFile>> {
-    if !files_from_used {
-        return Ok(None);
-    }
-
-    let mut file = NamedTempFile::new()?;
-    {
-        let writer = file.as_file_mut();
-        for entry in entries {
-            write_file_list_entry(writer, entry.as_os_str())?;
-            if zero_terminated {
-                writer.write_all(&[0])?;
-            } else {
-                writer.write_all(b"\n")?;
-            }
-        }
-        writer.flush()?;
-    }
-
-    Ok(Some(file))
-}
-
-fn write_file_list_entry<W: Write>(writer: &mut W, value: &OsStr) -> io::Result<()> {
-    #[cfg(unix)]
-    {
-        writer.write_all(value.as_bytes())
-    }
-
-    #[cfg(not(unix))]
-    {
-        writer.write_all(value.to_string_lossy().as_bytes())
-    }
-}
-
-fn push_toggle(args: &mut Vec<OsString>, enable: &str, disable: &str, setting: Option<bool>) {
-    match setting {
-        Some(true) => args.push(OsString::from(enable)),
-        Some(false) => args.push(OsString::from(disable)),
-        None => {}
-    }
 }
 
 fn push_file_list_entry(bytes: &[u8], entries: &mut Vec<OsString>) {
