+++ crates/core/Cargo.toml
@@ -19,6 +19,7 @@ rsync-compress = { path = "../compress" }
 libc = "0.2"
 base64 = "0.22"
 rsync-checksums = { path = "../checksums" }
+tempfile = "3.10"
 
 [features]
 default = []
